ref: update PR according to comments
