+++ src/sentry/static/sentry/app/components/forms/searchInput.tsx
@@ -10,7 +10,6 @@ type Props = {
   onChange: ({target: {value: string}}) => void;
   placeholder: string;
   smaller?: boolean;
-  width?: string;
 };
 
 class SearchInput extends React.Component<Props> {
@@ -19,9 +18,9 @@ class SearchInput extends React.Component<Props> {
   };
 
   render() {
-    const {placeholder, value, onChange, smaller, width} = this.props;
+    const {placeholder, value, onChange, smaller} = this.props;
     return (
-      <SearchWrapper smaller={!!smaller} width={width || ''}>
+      <SearchWrapper smaller={!!smaller}>
         <SearchIcon />
         <SearchField placeholder={placeholder} value={value || ''} onChange={onChange} />
         {value && value.length > 0 && (
@@ -57,10 +56,9 @@ const SearchReset = styled(props => <InlineSvg src="icon-circle-close" {...props
   }
 `;
 
-const SearchWrapper = styled('div')<{smaller: boolean; width: string}>`
+const SearchWrapper = styled('div')<{smaller: boolean}>`
   position: relative;
   display: inline-block;
-  width: ${p => (p.width ? p.width : 'auto')};
   ${SearchIcon} {
     top: ${p => (p.smaller ? '7px' : '11px')};
   }
