+++ src/sentry/static/sentry/app/views/organizationIntegrations/integrationViewController.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Control, {OrganizationIntegrations} from './control';
-import Test from './test';
+import Control, {OrganizationIntegrations} from './index';
+import Test from './integrationListDirectory';
 
 export default function IntegrationListView(props) {
   if (localStorage.getItem('USE_INTEGRATION_DIRECTORY') === '1') {
